+++ project_cylon/behook.py
@@ -4,8 +4,9 @@ from behave import configuration
 from behave import __main__
 
 ## add my wanted option to parser.
-configuration.parser.add_argument('--browser', help="Specify browser to run, firefox or chrome")
-configuration.parser.add_argument('--site', help="Specify site to run")
+configuration.parser.add_argument('--browser')
+configuration.parser.add_argument('--site')
+configuration.parser.add_argument('--debug', action='store_true')
 
 ## command that run behave.
 def main():